Add node deletion
This was much harder than it looked. The seamless-immutable library was not very seamless.

Mainly the problem was in not knowing whether a method was part of the object or in the Immutable
namespace. I'm going to take some time to get rid of that distinction as much as possible.